2017 MVAs now derived from same base class
